Complete Cart function
Full Cart methods done

+++ abcRestaurant/src/main/java/com/abcRestaurant/abcRestaurant/Model/Cart.java
@@ -18,5 +18,5 @@ public class Cart {
     private ObjectId id;
     private String userId;
     private Map<String, Integer> productId;
-    private double totalAmount;
+    private float totalAmount;
 }
